backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["gok@microsoft.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
semantic-kernel = "^1.3.0"
beautifulsoup4 = "^4.12.3"
azure-search-documents = "^11.4.0"
azure-ai-documentintelligence = "^1.0.0b4"
azure-identity = "^1.19.0"
openai = "^1.52.0"
tiktoken = "^0.8.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pyyaml = "^6.0.1"

ruff = "^0.8.0"
black = "^24.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
